+Gameloop fix +walls animation removed

+++ src/main/java/gameDA/config/output/Camera.java
@@ -8,9 +8,9 @@ import java.awt.image.BufferedImage;
 
 
 public class Camera {
-    private float x,y;
+    private int x,y;
 
-    public Camera(float x,float y){
+    public Camera(int x,int y){
         this.x = x;
         this.y = y;
 
@@ -26,20 +26,20 @@ public class Camera {
         if(y >= 1158) y = 1158;
 
     }
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public Camera setX(float x) {
+    public Camera setX(int x) {
         this.x = x;
         return this;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 
-    public Camera setY(float y) {
+    public Camera setY(int y) {
         this.y = y;
         return this;
     }
